refactor: 폴더 구조 변경, displayedAt 함수 변경, 댓글 스타일 변경
