+++ src/components/watch/Comments.tsx
@@ -1,8 +1,8 @@
 import { useEffect, useState } from "react";
 import styled from "styled-components";
-import { getComments } from "../api/api";
-import { commentsDummyData } from "../data";
-import type { CommentsProp, IComment } from "../types/commentsTypes";
+import { getComments } from "../../api/api";
+import { commentsDummyData } from "../../data/data";
+import type { CommentsProp, IComment } from "../../types/commentsTypes";
 import Comment from "./Comment";
 
 const Comments = ({ comments }: CommentsProp) => {
