+++ src/components/watch/MainVideo.tsx
@@ -1,7 +1,7 @@
 import styled from "styled-components";
-import { VideoDetailData } from "../types/videoDetailTypes";
-import { displayedAt } from "../utils/displayedAt";
-import { nFormatter } from "../utils/nFormatter";
+import { VideoDetailData } from "../../types/videoDetailTypes";
+import { displayedAt } from "../../utils/displayedAt";
+import { nFormatter } from "../../utils/nFormatter";
 
 interface MainVideoProps {
   videoDetailData: VideoDetailData;
@@ -28,11 +28,7 @@ const MainVideo = ({ videoDetailData }: MainVideoProps) => {
         {"조회수 "}
         {nFormatter(Number(videoDetailData?.items[0]?.statistics?.viewCount))}
         {"회 "} &nbsp;
-        {displayedAt(
-          new Date(
-            videoDetailData?.items[0]?.snippet?.publishedAt || "",
-          ).getTime(),
-        )}
+        {displayedAt(videoDetailData?.items[0]?.snippet?.publishedAt || "")}
       </ViewDateInfo>
       <ButtonWrapper>
         <LikeContainer>
