+++ src/components/watch/RelatedCard.tsx
@@ -1,22 +1,21 @@
 import React, { useState, useEffect } from "react";
-import { instance } from "../api";
+import { instance } from "../../api";
 import styled from "styled-components";
 import { Link } from "react-router-dom";
-import { displayedAt } from "../utils/displayedAt";
-import { nFormatter } from "../utils/nFormatter";
-import { videoTime } from "../utils/videoTime";
+import { displayedAt } from "../../utils/displayedAt";
+import { nFormatter } from "../../utils/nFormatter";
+import { videoTime } from "../../utils/videoTime";
 import { AiOutlineClockCircle } from "react-icons/ai";
 import { BiListCheck, BiDotsVerticalRounded } from "react-icons/bi";
-import { getVideoDetail } from "../api/api"
-import type {DataType, Video} from "../types/relatedTypes"
+import { getVideoDetail } from "../../api/api";
+import type { DataType, Video } from "../../types/relatedTypes";
 
 type Props = { item: DataType };
 
 function RelatedCard({ item }: Props) {
-  const videoDate: number = new Date(item.snippet.publishTime).getTime();
-  const [videoResult, setVideoResult] = useState<Video|null>(null);
+  const [videoResult, setVideoResult] = useState<Video | null>(null);
   const [isHovering, setIsHovering] = useState<boolean>(false);
-  const [isError, setIsError] = useState<string>("")
+  const [isError, setIsError] = useState<string>("");
   // const getDetail = async (videoId: string = item.id.videoId) => {
   //   const res = await instance.get(
   //     `/videos?part=snippet&part=contentDetails&part=player&part=statistics&id=${videoId}`,
@@ -26,7 +25,7 @@ function RelatedCard({ item }: Props) {
   useEffect(() => {
     getVideoDetail(item.id.videoId, setVideoResult, setIsError);
   }, [item]);
-  console.log(JSON.stringify(item))
+  console.log(JSON.stringify(item));
   const watchLink = `/watch/${item.id.videoId}`;
   return (
     <Link to={watchLink}>
@@ -37,7 +36,7 @@ function RelatedCard({ item }: Props) {
         <PreviewBox>
           <Thumbnail src={item.snippet.thumbnails.medium.url} />
           <VideoTime>
-            {videoTime(videoResult?.items[0]?.contentDetails?.duration||"")}
+            {videoTime(videoResult?.items[0]?.contentDetails?.duration || "")}
           </VideoTime>
           {isHovering ? (
             <HoverBox>
@@ -56,10 +55,11 @@ function RelatedCard({ item }: Props) {
           <RelatedChannel>{item.snippet.channelTitle}</RelatedChannel>
           <InfoBox>
             <p>
-              조회수 {nFormatter(Number(videoResult?.items[0]?.statistics?.viewCount))}
+              조회수{" "}
+              {nFormatter(Number(videoResult?.items[0]?.statistics?.viewCount))}
             </p>
             <InfoDot> • </InfoDot>
-            <p>{displayedAt(videoDate)}</p>
+            <p>{displayedAt(item.snippet.publishTime)}</p>
           </InfoBox>
         </DetailBox>
       </VideoCard>
