+++ src/components/watch/RelatedVideo.tsx
@@ -1,15 +1,15 @@
 import { useEffect, useState } from "react";
 import RelatedCard from "./RelatedCard";
 import styled from "styled-components";
-import { relatedVideoDummyData } from "../data";
-import type { IData } from "../types/relatedTypes"
+import { relatedVideoDummyData } from "../../data/data";
+import type { IData } from "../../types/relatedTypes";
 
 type Props = { relatedData: IData };
 
 function RelatedVideo({ relatedData }: Props) {
   return (
     <RelatedList>
-      {relatedData?.map((item:any) => (
+      {relatedData?.map((item: any) => (
         <RelatedCard key={item.id.videoId} item={item} />
       ))}
     </RelatedList>
