+++ src/pages/Home/index.tsx
@@ -2,7 +2,7 @@ import { instance } from "../../api/index";
 import { useEffect, useState } from "react";
 import styled from "styled-components";
 import Card from "../../components/Card";
-import { mainVideoDummyData } from "../../data";
+import { mainVideoDummyData } from "../../data/data";
 
 export interface VideoData {
   kind: string;
